+++ frontend/src/app/signup-page/signup-page.component.ts
@@ -22,15 +22,15 @@ interface Response {
 })
 export class SignupPageComponent implements OnInit {
   constructor(private http: HttpClient, private backendLink: BackendLinkService) {
-    const usrStr = localStorage.getItem('newUser');
-    if (typeof usrStr === 'string') {
-      this.newUser = JSON.parse(usrStr);
-    } else {
-      this.newUser = new User();
-    }
+    // const usrStr = localStorage.getItem('newUser');
+    // if (typeof usrStr === 'string') {
+    //   this.newUser = JSON.parse(usrStr);
+    // } else {
+    //   this.newUser = new User();
+    // }
   }
 
-  private newUser: User;
+  private newUser!: User;
   
 
   //// Validators
@@ -57,9 +57,7 @@ export class SignupPageComponent implements OnInit {
   }
 
   displayArrow: boolean = true;
-  hidePassword = true;
 
-  passwordMinLength: number = 10;
 
   // display down arrow if the user has not scrolled to the bottom of the page
   @HostListener('window:scroll', ['$event'])
@@ -73,64 +71,16 @@ export class SignupPageComponent implements OnInit {
   }
 
 
-  // To display the list of countries
-  countries: string[]= jsonFile.listOfCountries.map(country => country.name);
-  filteredCountries!: Observable<string[]>;
+
+  ngOnInit(): void {}
+
   
-
-  signupForm!: FormGroup;
-
-
-  ngOnInit(): void {
-    this.signupForm = new FormGroup({
-      username: new FormControl('', {
-        validators: [Validators.required]
-      }),
-      password: new FormControl('', {
-        validators: [Validators.required, Validators.minLength(this.passwordMinLength)]
-      }),
-      firstname: new FormControl('', {
-        validators: [Validators.required]
-      }),
-      surname: new FormControl('', {
-        validators: [Validators.required]
-      }),
-
-      street: new FormControl('', {
-        validators: [Validators.required]
-      }),
-      country : new FormControl('', {
-        validators: [this.countryValidator(this.countries), Validators.required] 
-      }),
-      postalCode: new FormControl('', {
-        validators: [Validators.required, this.postalCodeValidator()]
-      }),
-      city: new FormControl('', {
-        validators: [Validators.required]
-      })
-    });
-
-    
-    this.filteredCountries = this.signupForm.controls['country'].valueChanges.pipe(
-      startWith(''),
-      map(value => this._filter(value || '')),
-    );
-
-    SignupPageComponent.fillForm(this.newUser, this.signupForm)
-    
-  }
-
-  private _filter(value: string): string[] {
-    const filterValue = value.toLowerCase();
-
-    return this.countries.filter(country => country.toLowerCase().includes(filterValue));
-  }
 
   onSubmit(form: FormGroup) {
     //this.newUser = this.formFieldsToObject(form);
     //console.log(this.newUser);
-    this.newUser = SignupPageComponent.formGroupToUserObject(form);
-    localStorage.setItem('newUser', JSON.stringify(this.newUser));
+    //this.newUser = SignupPageComponent.formGroupToUserObject(form);
+    //localStorage.setItem('newUser', JSON.stringify(this.newUser));
     return this.http
       .post<Response>(this.backendLink.getRegisterUrl(), this.newUser).subscribe({
         next: (data: Response) => {
@@ -142,35 +92,6 @@ export class SignupPageComponent implements OnInit {
         }
       });
   }
-
-  private static formGroupToUserObject(form: FormGroup): User {
-    const value = form.value;
-    return {
-      username: value.username,
-      firstname: value.firstname,
-      surname: value.surname,
-      password: value.password,
-      address: {
-        street: value.street,
-        country: value.country,
-        city: value.city,
-        postalCode: value.postalCode 
-      }
-    }
-  }
-
-  private static fillForm(user: User, form: FormGroup): void {
-    form.setValue({
-      username: user.username,
-      firstname: user.firstname,
-      surname: user.surname,
-      password: user.password,
-      street: user.address.street,
-      country: user.address.country,
-      city: user.address.city,
-      postalCode: user.address.postalCode
-    });
-  } 
 
 }
